Updated copyright header message on mbed1768_asm.S
Give credit to ARM Limited for writing the assembly language code which was
extracted from the mbed SDK and switch to the MIT license as preferred by the
mbed team.  I have also added a comment indicating which portions of this
source file contain material that I added/modified.

Obtained permission from Simon Ford of the mbed team on May 14, 2012 via e-mail
to include the source code for these two routines in the mbed sources.