Allow GDB to connect on Ubuntu immediately after enumeration.
On Ubuntu, I see a AT+GCAP string get sent to the debugger serial port right
after enumeration and before GDB ever attempts to connect.  This ends up
setting the baud rate to the wrong value since MRI uses the LPC176x's auto
baud feature to detect the baud rate based on the initial AT+... string from
Ubuntu and not the '+' character first sent from GDB on connection.

This change modifies MRI to eat packets that are sent when first waiting for
GDB to connect if they don't start with the '+' character.  The packets are
eaten by just pulling all characters from the serial port until a delay of
at least 10 milliseconds is seen.  The SysTick timer on the Cortex-M3 is used to
perform this timing.  When the 10 milliseconds have elapsed, then the UART is
setup again to wait for the initial GDB connection character.