Include NULL terminator in filename length for Fopen/Funlink calls to gdb.
I was encountering an issue where on Linux, the files created by Fopen calls
issued from  MRI had garbage characters at the end.  This was caused by the
fact that the mbed semihost open/remove calls don't include the NULL terminator
in the filename length but looking at GDB's remote_fileio.c, I saw that it did
expect the NULL terminator to be included in this length.