MRI could hang on connect attempt.
I was seeing this hang more prominately when MRI wasn't configured for break on
init and then you later connected GDB to break into the running program.  It
turns out that the characters wouldn't come in quickly enough in this scenario
and some code that I added to work around an issue with Linux hosts would cause
the code to wait forever for the beginning of GDB's connect attempt.  I fixed
it by reordering some code so that a single '+' character received at the
beginning of a connect attempt wouldn't be thrown away and the debuggee
started back up inappropriately.