Clear dbg monitor pending bit on continue
When continuing execution of the debuggee from a fault, it is possible
that there is also a debug monitor interrupt pending due to the fact
that the instruction which hit the hard fault was executed as a single
step.  Now MRI clears this pending bit on exit so that it doesn't just
jump back into the debugger immediately for the same instruction.

I also updated a spelling mistake in a comment in some unrelated code.