Force double word stack alignment bit off in PSR
This change allows the debugger to change the SP value to an arbitrary
word aligned address and the registers will be properly placed on the
stack where the Cortex processor can pull them off during exception
return.  The previous code would do the wrong thing if the desired SP
value was double word aligned but the PSR stack align bit was set.