Catch memory faults when dereferencing PC
Previously the MRI debug monitor would generate memory faults when
the current PC was pointing to invalid memory and it tried to access
the contents of the current instruction via:
  getFirstHalfWordOfCurrentInstruction()
  getSecondHalfWordOfCurrentInstruction()
The code now throws an exception in these cases and code was added to
catch this exception and default the behaviour that best makes sense
when the PC isn't valid:
  For example Platform_AdvanceProgramCounterToNextInstruction() does
  nothing if the PC is already pointing to invalid memory.

Once this fix was made, the pinkysim BL tests which use the maximum
negative and positive offsets now function as expected.  Previously
they would leave the system in a state where it though good memory
access requests had generated a fault when it was actually the internal
reading of the current instruction (at an invalid address) that
generated the fault.