Enable single stepping over SVC instructions
Previously single stepping over a SVC instruction would result in a
hard fault because MRI would issue the single step with the priority
boosted to a level which wouldn't allow the SVC handler to run.

The code now detects this scenario and just sets a hardware breakpoint
on the SVC_Handler entry point and continues execution instead of
single stepping.  This does mean that other interrupt handlers will
get a chance to run as well but in most cases the user will never
notice this unless they have breakpoints set in those other handlers.
I tests this code out and GDB can now step into a SVC handler and back
out with no issues.