Add back in erroneously removed memory tests
I removed failing tests during the last two commits because I thought
those cases were failing because the mentioned fault when writing the
final result to memory was no longer valid.  This wasn't the case and
removing these test cases reduced the amount of code coverage in
core/memory.c.  I have added back these tests and correct the ordinal
of the allocation attempt to fail so that the tests now exercise the
correct code again.