Drop 200 bytes from debugger stack
I had previously reserved 512 bytes for the debugger stack.  This
commit trims it down to 312 bytes.

I ran a scenario which included file I/O commands and this used up a
maximum stack of 272 bytes. Setting to 312 will give 40 bytes of head
room for future growth.

New MRI memory usage:
   text	   data	    bss	    dec	    hex	filename
   9004	    116	    552	   9672	   25c8	LPC1768/Blink.elf
  23100	    116	   1204	  24420	   5f64	LPC1768/Blink.elf
  -----	-------	-------
  14096       0     652

A rough breakdown of the 652 bytes of bss still used:
    137 bytes for packet buffer
     68 bytes for register context
    312 bytes for debugger stack
     40 bytes of other data to restore on continue like MPU settings,
        base priority, MSP, etc.
     36 bytes of core MRI state (Packet and Buffer structure + a few
        tracking words.)
     40 bytes for mbed UUID (MAC address) and 1 flag word.
      8 bytes for LPC1768 state.
    ---
    641 bytes + alignment padding + exception code