Minimum FPU support
This commit will properly adjust the SP if FPU register auto-stacking
has taken place during exception entry.  With this commit, MRI should
be able to properly debug a device with the FPU enabled but it still
doesn't give the programmer access to the FPU registers themselves but
it can properly enter and leave the debug monitor without corrupting
the state of the processor.