T packet now returns R7 contents instead of R12
R7 is actually the frame pointer for Thumb2 so it is better to return
it in the short T context dump instead of R12.

This makes source level single stepping more efficient when the
frame pointer is used by the compiler since GDB will request the full
context to get the R7 (FP) register contents on each machine code
single step return otherwise.