Add 'C' and 'S' GDB command support
MRI already supported the 'c' and 's' commands for continuing execution
and single stepping. These new commands are similar except they take a
signal argument as well which MRI ignores. Previously these commands
weren't required but the 7.12.1 release of GDB sometimes fails to
fallback to the 's'/'c' versions of the commands and execution isn't
resumed.