Move headers from include/ to core/
This makes the file layout more compatible with other build systems
used when building MRI directly in other projects. Rather than adding
core/ to the INCLUDE path, the root of the MRI project is instead
added and the core/ is explicitly prepended to the header names when

I updated the copyright messages to use my github account URL instead
of the mbed URL.

I also use #include <> and #include "" more consistently than I did
previously.

Note: This commit also includes the removal of a lot of trailing
      whitespace that my old editor setup used to leave around in my
      source files.