Remove use of MPU to disable caching
I removed the code in MRI which uses the MPU to disable caching.
Disabling caching on a Cortex-M7 caused the core to see stale data
when the newer data was still sitting in the cache. I initially used
the MPU this way on Cortex-M3/M4 cores so that bus faults for writes
to bad addresses would be precise and properly reported back to GDB
when the GDB user attempted to write to an invalid piece of memory.
After removing this MPU code, I found that MRI still catches invalid
writes and reports them back to GDB on LPC1768 targets where I first
encountered the problem and wrote this code. My memory write code path
in MRI must now take long enough for the fault to be caught before I
make the check.