Expose mriFaultHandler as global symbol
I modified armv7-m_asm.S to expose mriFaultHandler as a global symbol
named __mriFaultHandler so that HardFault_IRQn,
MemoryManagement_IRQn, and BusFault_IRQn could be dynamically set to it
when initializing the MRI library. This makes it possible to use the
MRI library on platforms which have their own handlers defined in the
FLASH based interrupt vector that stop us from placing the MRI ones
there.

The MRI implementations of the fault handlers may still need to be
deleted from armv7-m_asm.S if the build system picks up the MRI
versions erroneously or the build won't tolerate multiply defined
symbols.