Default MRI_DEVICE_HAS_FPU based on architecture
The MRI build system will make sure that MRI_DEVICE_HAS_FPU is set
correctly for each platform it supports. In some other build systems,
MRI_DEVICE_HAS_FPU won't be passed in on compiler's command line so use
the target Cortex-M type to determine if it has a FPU or not.