Add __mriCore_SetTempBreakpoint() to MRI
This commit adds the ability to add a single temporary breakpoint that
when hit, passes control to the caller specified callback. This
callback can return 1 if it wants to resume execution after handling
this breakpoint or 0 if it wants to halt in GDB instead. I added this
API to make it possible to break in main() or other functions of
interest and maybe even run special code when the breakpoint hits.

Other changes include:
* I renamed the old Platform_SetHardwareBreakpoint() function to
  Platform_SetHardwareBreakpointOfGdbKind() and now have the function
  named Platform_SetHardwareBreakpoint() just take an address and it
  will figure out if the instruction at that address is 16-bit or
  32-bit. The original function gets this information passed in via
  GDB's kind parameter. I did the same for
  Platform_ClearHardwareBreakpoint() as well.
* I updated the SVC single stepping code which set breakpoints so that
  it now takes advantage of the new breakpoint API.
* MRI previously assumed that the DebugMon interrupt was running at
  priority level 0 when it elevates the priority of the interrupted
  code while single stepping. I modified the code to now read the
  actual priority setting of the DebugMon interrupt instead of assuming
  its value.
* I added unit tests to cover the new temp breakpoint functionality.