Remove "hwbreak" & "swbreak" support
I decided to not return these stop reasons in T packets because GDB is
able to work out when a breakpoint is hit without them and returning
"swbreak" on hardcoded breakpoints sometimes confused GDB even though
the GDB remote protocol spec says that it shouldn't.

"monitor reset" wouldn't hit the initial hardcoded breakpoint after
reset when "swbreak" was given as a reason. GDB would just silently
ignore it and tell the MRI stub to continue execution over it.