Implement qf/sThreadInfo request to list threads to GDB
If a port implements Platform_RtosGetThreadCount() and
Platform_RtosGetThreadArray() then MRI will respond to GDB's
qfThreadInfo & qsThreadInfo requests with a list of the IDs of the
RTOS threads. The code is currently setup so that a 0 count returned
from Platform_RtosGetThreadCount() will have MRI return an empty
response to these requests, indicating it doesn't support these
type of requests.

I also extended platformMock_CommInitReceiveData() and
platformMock_CommInitReceiveChecksummedData() to accept up to 3
separate packets as parameters instead of the previous 2. This allowed
for new unit tests to send "qfThreadInfo", "qsThreadInfo", and a "c"
(continue) request.