Added qThreadExtraInfo support
This allows GDB to call and retrieve extra information about a
particular thread ID as a string. This would typically be used for
returning information such as thread state (running, waiting, etc)
and/or the name of the thread. The platform port provides an
implementation of the Platform_RtosGetExtraThreadInfo() API to provide
the NULL terminated string to be returned to GDB. The MRI core library
provides a weak implementation that just returns a NULL pointer.

I also added a Buffer_WriteStringAsHex() function to output a ASCII
based string in hexadecimal format. This is how GDB expects the thread
extra info to be returned. I updated the
writeStringToExclusiveGdbCommChannel() in gdb_console.c to make use of
this new API as well.