Refactor register context handling code
I renamed the ScatterGather class to MriContext and now use it from
the higher level MRI core code to access the CPU register contents.

Ports no longer need to implement the Platform_CopyContextToBuffer()
and Platform_CopyContextFromBuffer() APIs and instead the halting
thread's context is passed into mriDebugException() on halt and used
when GDB requests to read or write the registers.

This change was made in preparation for a future commit which will
allow switching contexts between various threads with the Hg command.