Add support for Hg command
GDB sends this command to indicate which thread's registers should be
reference by the next 'g' or 'G' command. Ports can now implement the
Platform_RtosGetThreadContext() API to switch the register context to
the specified thread ID and the MRI core will use the returned context
for future register read and write commands.