Added support for the T command
The T command is sent by GDB to see if a particular thread ID is still
active or not. A port need to implement a Platform_RtosIsThreadActive()
function for this to work. GDB sends this command to a remote stub when
the user types the "thread id#" command into the user interface to see
if the specified ID is valid or not.