Update vCont to handle specific thread id
vCont can now handle specific thread ids as long as they match the id
of the thread which caused the halt to occur. Any other id will still
generate an error. In the normal cases I have seen so far, this is the
id that GDB uses when single stepping a multi-threaded program.