Add thread state support for continue,step,detach
Usually GDB will use vCont for continue and single stepping if threads
are supported but it is probably best to have the simpler continue ('c'
and 'C') and step ('s' and 'S') GDB command handler makes calls to
Platform_RtosSetThreadState() if it is supported. The detach handler
needs to call Platform_RtosSetThreadState() as well.