Initial nRF52-DK support
This commit makes it possible to use MRI to debug the nRF52 Development
Kit over the virtual com port. A lower (non zero) priority level will
have to be passed into mriInit() via a "MRI_PRIORITY=xx" parameter
string if the BLE stack is running since it will fault if it gets
stopped by MRI.

The Makefile underwent a bunch of changes to enable building all
sources for no FPU support, FPU using the softfp ABI, and FPU using
the hard ABI. I build 2 versions of the nRF52 library, one for each
type of FPU ABI.

When using the Nordic SDK for building your binaries, you will need
to copy devices/nrf52/nrf52_asm.S into your project and build it
there as the linker won't pull in the UART ISR handler from the MRI
library, only from a .o file that is a part of the project.

NOTE: To use the virtual COM port provided by the nRF52DK for
      communicating with MRI, you should be aware of 2 things:
      * You need to disable the USB mass storage device interface so
        that the VCOM interface can accept packets larger than 64 bytes
        from GDB. See following link for more info:
        https://devzone.nordicsemi.com/nordic/nordic-blog/b/blog/posts/nrf-connect-v10-release#osxissue
      * If you use nrfjprog to program the device then press the reset
        button to manually reset the nRF52 uC after programming to make
        sure that the interface chip shuts down the SWD hardware debug
        connection which will interfere with MRI's use of monitor mode.