Move UARTE0_UART0_IRQHandler into nrf52_uart.c
This means that it will get pulled into a project using the nRF52
version of the MRI library with no special build tweaks. If any nRF52
UART code from nrf52_uart.c is needed to build a project then this
strong implementation of UARTE0_UART0_IRQHandler() will be pulled in
too.