Cleanup recent packet escaping commit
Placed newly added static functions to core/packet.c in the typical
order used by MRI code base.

Also fixed the handling of a truncated packet in the middle of an
escape sequence. One of the unit tests failed without this fix as it
verified that the code properly handled this situation which it no
longer did after the unescaping code was moved from memory.c