Use unique bkpt #'s for each newlib stub
Previously all of the newlib assembly language stubs just contained a
"bkpt 0xff" and MRI used the program counter to figure out which one
had been hit. This works well when MRI is linked directly into the
debuggee but it didn't work when MRI was in a separate bootloader like
mriblue_boot. This commit gives each newlib stub its own unique bkpt
number as seen in newlib_stubs.h

mriSetDebuggerHooks() is now implemented as a newlib stub so that it
too can be called from a debuggee when MRI is running in a separately
linked bootloader.