Add Platform_CommSendBuffer()
This new API can be implemented for particular ports so that a whole
packet can be sent at once rather than byte by byte as needs to be done
when using the existing Platform_CommSendChar() routine. This is useful
when using something like USB or TCP/IP for sending data back to GDB.

There is a default weak implementation of Platform_CommSendBuffer() in
packet.c which will just call Platform_CommSendChar() for each byte to
be sent if a port hasn't provided their own strong implementation. This
default keeps existing ports working.

I switched the porting documentation over to MarkDown instead of creole
and then made some updates related to the addition of
Platform_CommSendBuffer() and other changes in the API since the
documentation was last touched.

The Packet structure now owns 2 Buffer objects. packetBuffer contains
the whole buffer needed for the packet, including the 4 bytes of
overhead for '$', '#', and hex checkdum digits. dataBuffer starts 1
byte after packetBuffer to reserve room for the '$' packet header byte.
dataBuffer is also initialized as being 4-bytes shorter than
packetBuffer. I removed the buffer field from the MriCore structure as
it now lives on within the Packet object itself alongside packetBuffer.

I added a Packet_Init() method to initialize the Packet object with
backing storage for its 2 Buffer objects (packetBuffer and dataBuffer).
The CortexMState::packetBuffer array was increased by 4 bytes to
accomodate the extra packet overhead that now needs to be placed in
this buffer. Similar changes needed to be made to various unit tests.

NOTE: I ended up removing the code which would automatically escape
      bytes being sent to GDB added recently by @xiaoxiang781216. This
      was done because it is much harder to expand the bytes found in
      Packet::dataBuffer in place than just expand as sending each byte
      over to GDB. So far there are no commands supported by MRI that
      require outgoing data to be escaped anyway. If we need this
      functionality in the future then it would probably make sense
      to add a method to Buffer which escapes the bytes as they are
      placed into Packet::dataBuffer by the command handler.