Allow single stepping when interrupts are disabled
NOTE: The changes below to enable debugging while PRIMASK is set to 1
      only take effect when MRI is configured to run at priority level
      0. There is no safe way to facilitate this when running at lower
      priorities since temporarily setting PRIMASK to 0 to allow the
      DebugMon interrupt to fire would also allow other higher priority
      interrupts to execute and it may be these interrupts that the
      user code was actually trying to disable when they entered their
      critical section.

When MRI is running at interrupt priority level 0, no BASEPRI setting
can disable it but setting PRIMASK will. If you encounter a breakpoint
while PRIMASK is set then the debug event will be escalated to a
HardFault. This commit includes a change which saves off the current
BASEPRI and PRIMASK settings when such a HardFault occurs and then
clears PRIMASK before pending a DebugMon interrupt. MRI will restore
BASEPRI and PRIMASK to their original settings in the context returned
to GDB and these will be restored to the CPU when execution is resumed.
I also added code which updates the recorded fault state enough that
MRI will tell GDB that the cause was a a debug event and not a hard
fault.

The above change deals with the situation where you hit a breakpoint
when interrupts are disabled but if you try to single step through such
a critical section when MRI is running at priority level 0 then the
single stepping trap won't fire until the CPU has left the critical
section created via PRIMASK. This commit also includes changes to
Platform_EnableSingleStep() which recognize if you are stepping over
instructions that deal with PRIMASK or BASEPRI and simulates their
effects on the register context (MSR, CPS, and MRS) while keeping
the actual PRIMASK set to 0 and BASEPRI set to a priority level of
1 so that all other interrupts other than MRI are disabled while
single stepping.

When MRI is next re-entered after it has massaged the PRIMASK and
BASEPRI to facilitate handling of debug events which get escalated to a
Hard Fault or to disable other interrupts while single stepping, it
will restore the values of PRIMASK and BASEPRI from mriCortexMState
and place them in the register context so that GDB sees the correct
values.

Other changes in this commit include:
* Now restores PRIMASK, FAULTMASK, and CONTROL from MRI's register
  context into the CPU's actual registers. This was done in the
  mriExceptionHandler assembly language routine.
* Updates mocks and tests to be able to verify the changes I made to
  cmd_step.c as part of this commit.