Don't reset on single step
Previously if a "monitor reset" command had been sent to MRI, it would
reset the device on a single step or a continue command. When GDB
continues from a halt at a breakpoint, it will first single step over
the instruction and then issue the continue command to MRI. In certain
scenarios this would lead to issues if the CPU reset occurred on the
single step since it would then not stop at the expected location. Now
MRI only issues the CPU reset on continue commands.

I also deleted a duplicate line that snuck in when I converted the line
endings 3 years ago.