Handle faults during ranged single stepping
Fixes issue #35.

Previously hitting a crash in the middle of a single step range would
lead to a hang. It now handles SIGSEGV, SIGBUS, and SIGILL like it
already handled SIGINT (CTRL+C) by ignoring that it was single stepping
and just handle the higher priority event that has occurred instead.

I also updated the version build date and added a note to the README
about the OS ABI version of GDB that should be used with MRI.