Make cache code more generic
The goal of this commit is to make the code which syncs the I cache to
the D cache more generic so that it can run on all ARMv7-M compatible
microcontrollers. It will detect when running on Cortex-M
microcontrollers without caches and it will just return. On current
Cortex-M7 microcontrollers with caches, it will synchronize the
I cache just like the previous code contributed by @icecream95. The
only major difference is that it only synchronizes to the point of
unification instead of coherency. It will also detect at runtime when
run on future microcontrollers with cache lines that are not 32-bytes in
length and work on those as well.