Truncate long writes to stdout
Previously long writes to stdout during semihosting would just be
dropped. They are now truncated and the number of bytes successfully
written are returned. This allows the truncation to be detected by the
standard library and it will send down another write with the rest of
the data so the user sees everything they were meant to see in the GDB
console.

* Buffer_Write*StringAsHex() routines always truncated the writes so
  that it didn't overflow the buffer but had a void return type. They
  now return the number of bytes successfully written to the buffer.
* WriteSizedStringToGdbConsole() routines now return the number of
  bytes successfully sent to GDB as well. Before it would detect the
  truncation and not send the packet to GDB at all.