Add MRI_ALWAYS_USE_HARDWARE_BREAKPOINT support
Setting MRI_ALWAYS_USE_HARDWARE_BREAKPOINT macro to non-zero value from
Makefile will force software breakpoints to be treated as hardware
breakpoints. This is done to support platforms that don't return memory
XML to indicate which memory regions are FLASH and which are RAM. This
change assumes that in such situations GDB is actually asking to set a
breakpoint in FLASH when sending the software breakpoint command.

This is an alternative fix for PR #39.