Clear MRI_FLAGS_RESET_ON_CONTINUE after reset
When MRI is running in monitor mode, the device running MRI reboots and
MRI will be re-initialized after calling Platform_ResetDevice(). When
run as a hardware debugger this isn't the case so I added the explicit
clear of MRI_FLAGS_RESET_ON_CONTINUE so that it doesn't keep resetting
the device on every subsequent 'continue' command.