Pass filename length into newlib stubs
The IssueGdbFile*() APIs need to know the length of the strings that
their pFilename parameters point to. Previously the
handleNewlibSemihost*Request() handlers would call mri_strlen() to get
these lengths and pass them into the IssueGdbFile*() functions. This
doesn't work for mri-swd where MRI is running on a different device and
can only access the target's memory over SWD.

This commit forces the call of the C based mriNewLib_Semihost*()
functions to pass in these filename lengths (including the '\0', null
terminator) and they will be passed through down to the IssueGdbFile*()
functions. The caller of those mriNewLib_Semihost*() functions run on
the target themselves so they have no problems making the required
strlen() call.