Allow MRI to work on 16-bit and 64-bit targets
Previously MRI was hardcoded for targets which used 32-bit values for
their registers and addresses. I am working on a project for the Z80
which needs 16-bit support and in the future I might want to add 64-bit
ARM targets as well. This commit gets rid of the hardcoded 32-bit types
used throughout the MRI code and replaces them with uintmri_t and
intmri_t types that can be set the correct types via the
MRI_UINT_TYPE and MRI_INT_TYPE macros.

Other changes include:
* Switch from uint32_t to size_t where it is even more appropriate than
  using the new uintmri_t type. An example would be the
  Platform_GetPacketBufferSize() and Platform_GetTargetXmlSize()
  functions.
* When reading/writing memory on the target inside of the core, there
  were places that the addresses were converted to void* but now they
  are maintained as uintmri_t.
* Removed the ADDR32_TO_POINTER macro that was previously used to
  generate 64-bit pointers from 32-bit addresses passed into the MRI
  core as they are no longer needed since the MRI core is now  built as
  64-bit when unit testing on a 64-bit host.
* I added support for reading/writing 64-bit values in memory.c
* Added Platform_MemRead64() and Platform_MemWrite64() APIs in
  platforms.h
* Updated the elements in PlatformWatchpointType to make them a
  bitmask. This made it easier to implement watchpoint support in my
  Z80 debugger.
* Updated build date in version.h
* Updated unit tests to work when uintmri_t is 64-bit as it is on my
  64-bit host.
* Updated CppUTest to work with my tests when they use 64-bit integers.